test refactoring WIP
