+++ app/concepts/challenge/views/list_detail.erb
@@ -1,7 +1,7 @@
 <li>
   <div class="pic">
     <%= link_to challenge_path(challenge) do %>
-      <%= concept(Image::Cell,challenge) %>
+      <%= concept(Image::Cell,challenge, current_participant: current_participant) %>
     <% end %>
   </div>
   <div class="details">
@@ -10,8 +10,8 @@
     <p class="status active"><%= remaining_text %></p>
   </div>
   <div class="stats">
-    <%= concept(Challenge::Cell::Statistic,challenge,type: 'submission') %>
-    <%= concept(Challenge::Cell::Statistic,challenge,type: 'participant') %>
-    <%= concept(Challenge::Cell::Statistic,challenge,type: 'view') %>
+    <%= concept(Challenge::Cell::Statistic,challenge,type: 'submission', current_participant: current_participant) %>
+    <%= concept(Challenge::Cell::Statistic,challenge,type: 'participant', current_participant: current_participant) %>
+    <%= concept(Challenge::Cell::Statistic,challenge,type: 'view', current_participant: current_participant) %>
   </div>
 </li>
