+++ app/concepts/challenge/views/show.erb
@@ -1,1 +1,1 @@
-<%= concept('challenge/cell/list_detail', collection: challenges) %>
+<%= concept(Challenge::Cell::ListDetail, collection: challenges, current_participant: current_participant) %>
