+++ app/concepts/template/cell.rb
@@ -1,7 +1,7 @@
 class Template::Cell < Cell::Concept
   include Pundit
-  include Devise::Controllers::Helpers
-  Devise::Controllers::Helpers.define_helpers(Devise::Mapping.new(:participant, {}))
+  #include Devise::Controllers::Helpers
+  #Devise::Controllers::Helpers.define_helpers(Devise::Mapping.new(:participant, {}))
   include Escaped
   include ActionView::Helpers::JavaScriptHelper
   include ActionView::Helpers::DateHelper
