+++ app/views/articles/edit.html.erb
@@ -1,1 +1,2 @@
-<%= render partial: 'articles/form/form', locals: { article: @article, title: 'Edit Tutorial'} %>
+<%= render partial: 'articles/form/form',
+           locals: { article: @article, title: 'Edit Tutorial'} %>
