+++ app/views/articles/new.html.erb
@@ -1,1 +1,2 @@
-<%= render partial: 'articles/form/form', locals: { article: @article, title: 'New Tutorial'} %>
+<%= render partial: 'articles/form/form',
+           locals: { article: @article, title: 'New Tutorial'} %>
