+++ app/views/challenges/_challenges.html.erb
@@ -1,1 +1,3 @@
-<%= concept(Challenge::Cell::ListDetail, collection: @challenges) %>
+<%= concept(Challenge::Cell::ListDetail,
+              collection: @challenges,
+              current_participant: current_participant) %>
