+++ app/views/challenges/edit.html.erb
@@ -1,1 +1,2 @@
-<%= render partial: 'challenges/form/form', locals: {challenge: @challenge, title: 'Edit Challenge'} %>
+<%= render partial: 'challenges/form/form',
+           locals: {challenge: @challenge, title: 'Edit Challenge'} %>
