+++ app/views/challenges/new.html.erb
@@ -1,1 +1,2 @@
-<%= render partial: 'challenges/form/form', locals: { challenge: @challenge, title: 'New Challenge' } %>
+<%= render partial: 'challenges/form/form',
+           locals: { challenge: @challenge, title: 'New Challenge' } %>
