+++ app/views/landing_page/index.html.erb
@@ -19,7 +19,7 @@
     </div>
 
     <ul class="list-challenges">
-      <%= concept(Challenge::Cell::ListDetail, collection: @challenges) %>
+      <%= concept(Challenge::Cell::ListDetail, collection: @challenges, current_participant: current_participant) %>
     </ul>
   </div>
 </content>
