+++ app/views/layouts/application.html.erb
@@ -40,10 +40,10 @@
 
   <body id='controller-action', data-controller="<%= controller.controller_name %>" data-action="<%= controller.action_name %>">
     <div data-behavior="progress-bar" class="progress-bar"></div>
-    <%= concept(Header::Cell) %>
+    <%= concept(Header::Cell, nil, current_participant: current_participant) %>
     <%= render 'layouts/flash_messages' %>
     <%= yield %>
-    <%= concept(Footer::Cell) %>
+    <%= concept(Footer::Cell, nil, current_participant: current_participant) %>
     <%= render partial: 'layouts/google_analytics_body' %>
   </body>
 </html>
