+++ app/views/leaderboards/_leaderboards.html.erb
@@ -1,1 +1,4 @@
-<%= concept(Leaderboard::Cell::TableRow, collection: @leaderboards, challenge: @challenge) %>
+<%= concept(Leaderboard::Cell::TableRow,
+              collection: @leaderboards,
+              challenge: @challenge,
+              current_participant: current_participant) %>
