+++ app/views/organizers/_form.html.erb
@@ -13,7 +13,10 @@
         <%= f.text_field :organizer %>
         <label>Profile Picture</label>
         <div class="upload">
-          <%= concept(Image::Cell,organizer,klass: 'img-thumb') %>
+          <%= concept(Image::Cell,
+                        organizer,
+                        klass: 'img-thumb',
+                        current_participant: current_participant) %>
           <%= f.file_field :image_file, class: 'file-upload' %>
           <% if organizer.image_file.present? %>
             <%= link_to 'Delete', organizer_remove_image_path(organizer) %>
