+++ app/views/organizers/edit.html.erb
@@ -1,1 +1,2 @@
-<%= render partial: 'organizers/form', locals: { organizer: @organizer, title: 'Edit Host' } %>
+<%= render partial: 'organizers/form',
+           locals: { organizer: @organizer, title: 'Edit Host' } %>
