+++ app/views/organizers/new.html.erb
@@ -1,1 +1,2 @@
-<%= render partial: 'organizers/form', locals: { organizer: @organizer, title: 'New Host' } %>
+<%= render partial: 'organizers/form',
+           locals: { organizer: @organizer, title: 'New Host' } %>
