+++ app/views/organizers/show.html.erb
@@ -1,1 +1,3 @@
-<%= concept(Organizer::Cell,@organizer) %>
+<%= concept(Organizer::Cell,
+              @organizer,
+              current_participant: current_participant) %>
