+++ app/views/topics/_topics.html.erb
@@ -1,1 +1,3 @@
-<%= concept(Discussion::Cell::ListDetail, collection: @topics) %>
+<%= concept(Discussion::Cell::ListDetail,
+              collection: @topics,
+              current_participant: current_participant) %>
