+++ app/views/topics/show.html.erb
@@ -1,5 +1,12 @@
 <content>
-  <%= concept(Challenge::Cell::ChallengeMasthead, @challenge) %>
-  <%= concept(Challenge::Cell::ChallengeSubnav, @challenge, tab: 'discussion') %>
-  <%= concept(Discussion::Cell, @topic) %>
+  <%= concept(Challenge::Cell::ChallengeMasthead,
+                @challenge,
+                current_participant: current_participant) %>
+  <%= concept(Challenge::Cell::ChallengeSubnav,
+                @challenge,
+                tab: 'discussion',
+                current_participant: current_participant) %>
+  <%= concept(Discussion::Cell,
+                @topic,
+                current_participant: current_participant) %>
 </content>
