+++ spec/features/participants/participant_deletes_account_spec.rb
@@ -1,0 +1,6 @@
+require "rails_helper"
+
+
+feature "participant deletes account" do
+  # TODO
+end
