+++ spec/features/participants/profile_fields_visibility_spec.rb
@@ -5,15 +5,14 @@ feature 'API key visibility', js: true do
   let(:admin) { create :participant, :admin }
 
   scenario "API key visible on participant's own profile" do
-    sign_in(participant)
+    log_in(participant)
     visit participant_path(participant.slug)
     expect(participant.api_key).not_to be_nil
     expect(page).to have_content participant.api_key
-    save_and_open_page
   end
 
   scenario "API key visible to admin users" do
-    sign_in(admin)
+    log_in(admin)
     visit participant_path(participant.slug)
     expect(participant.api_key).not_to be_nil
     expect(page).to have_content participant.api_key
@@ -31,14 +30,14 @@ feature 'Email field visibility', js: true do
   let(:admin) { create :participant, :admin }
 
   scenario "Email address is not visible on participant's own profile" do
-    sign_in(participant)
+    log_in(participant)
     visit participant_path(participant.slug)
     expect(participant.email).not_to be_nil
     expect(page).not_to have_content participant.email
   end
 
   scenario "Email key visible to admin users" do
-    sign_in(admin)
+    log_in(admin)
     visit participant_path(participant.slug)
     expect(page).to have_content participant.email
   end
